+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased](https://github.com/nationalarchives/tna-frontend/compare/v0.6.0...HEAD)
 
 ### Added
+
+- Capture user preferences (such as dark mode, reduced motion etc.) in GA4
+
 ### Changed
 
 - Ensure accordion `aria-label` attributes are trimmed